changed the protocol to exachange the data between client and host. Fixed issue with data loosing.
The protocol is simple - send text from STDOUT to socket as it is. If it has large size (more then 512 chars) it is splitted into chunks and chunks by chunks are sent.
Then receiver join all the chunks and sends whole buffer to clipboard on server side.
Tested for very large buffers - 2M works fine.